Basic restaurant api

+++ database/migrations/2019_02_16_180907_create_restaurants_table.php
@@ -18,7 +18,7 @@ class CreateRestaurantsTable extends Migration
             $table->string('name');
             $table->string('category');
             $table->unsignedTinyInteger('price');
-            $table->unsignedTinyInteger('rating');
+            $table->float('rating');
             $table->unsignedTinyInteger('processTime');
             $table->timestamps();
         });
